agentcore-chat/pyproject.toml
[project]
name = "agentcore-chat"
version = "0.1.0"
description = "Terminal-based conversational interface for Amazon Bedrock AgentCore"
readme = "README.md"
authors = [
    { name = "Brian Loyal", email = "bloyal@amazon.com" }
]
requires-python = ">=3.12"
dependencies = [
    "bedrock-agentcore",
    "bedrock-agentcore-starter-toolkit",
    "boto3",
    "click",
    "strands-agents",
    "strands-agents-tools",
]

[project.scripts]
chat = "agentcore_chat.chat:chat"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"